requirements.txt
aiofiles==23.1.0
aiogram==2.24
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
loguru==0.7.1
magic-filter==1.0.11
multidict==6.0.4
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
